requirements.txt
snowflake-snowpark-python==0.7.0
streamlit==1.11.1
plotly==5.9.0
pandas==1.4.3